orderCard componentizado separadamente, algumas pequenas alterações de layout.
